completed all actions

+++ src/module/todo/todo.repository.ts
@@ -1,35 +1,31 @@
 import { db } from "#prisma/db.js";
 
-type TodoData = Parameters<typeof db.orm.public.Todo.create>[0];
-type TodoStatus = NonNullable<TodoData["status"]>;
-
-type UpdateParams = {
-  id: string;
-  data: TodoData;
-};
-type FindByStatusParams = {
-  userId: string;
-  status: TodoStatus;
-};
+import type {
+  DeleteTodoParams,
+  FindByIdParams,
+  FindByStatusParams,
+  TodoData,
+  UpdateParams,
+} from "./type.js";
 
 const create = (data: TodoData) => {
   return db.orm.public.Todo.create(data);
 };
 
-const update = ({ id, data }: UpdateParams) => {
-  return db.orm.public.Todo.where({ id }).update(data);
+const update = ({ id, userId, data }: UpdateParams) => {
+  return db.orm.public.Todo.where({ id, userId }).update(data);
 };
 
-const remove = (id: string) => {
-  return db.orm.public.Todo.where({ id }).delete();
+const remove = ({ id, userId }: DeleteTodoParams) => {
+  return db.orm.public.Todo.where({ id, userId }).delete();
 };
 
-const findById = (id: string) => {
-  return db.orm.public.Todo.where({ id }).first();
+const findById = ({ id, userId }: FindByIdParams) => {
+  return db.orm.public.Todo.where({ id, userId }).first();
 };
 
 const findByStatus = ({ userId, status }: FindByStatusParams) => {
-  return db.orm.public.Todo.where({ userId, status });
+  return db.orm.public.Todo.where({ userId, status }).all();
 };
 
 const findAll = (userId: string) => {
